Support Updating Code During Employment
Erlang supports changing the code executed during run-time.
The changes can be applied without stopping the server. The
main process for listening on the specified port is now
started in a thread (which makes it possible to c(...) that
module). The code for handling a request is now in a separate
module, connection_handler.erl, and can be loaded separately
as well.

For now I have a PORT-constant defined in both erl-files. When
there are "config" files for the server (instead of abusing
the unix filesystem structure), this can be dropped.