Adjust Connection Handler to Read Config
Side note: It might be necessary to have a separate
configuration for the server. The server-config
would contain information like port, hostname, etc. .
The handler configuration would contain everything necessary
for a single connection. Since it is read every time a
connection is opened, it can be changed while the server is running;
new connections will automatically be handled with the updated config.